backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
redis==5.0.4
# python-jose removed[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
slowapi==0.1.9

# Structured logging
structlog==24.1.0
anyio[trio]==4.3.0

PyJWT==2.8.0
cryptography>=42.0.0
